run-kvm: collect HW info and logs from the host
As run-mmtests.sh runs inside guests, both logs and hardware and
configuration details are collected from within each guest. This is
fine, but we most of the time want/need similar info about the host as
well.

Add the capability to collect such logs to run-kvm.sh. This is done by
exporting the logging capability from run-mmtests.sh into common code,
and defining the proper paths and variable in run-kvm.sh too.

Note that we do not collect host logs & info by default, but only if the
`-L` option is provided to run-kvm.sh itself. This is because it may be
the case that we are not actually itnerested in host logs, or we are but
run-kvm.sh is driven by a more high level component which handles log
collection by itself.

If collected, the logs will be available inside the RUNNAME-host work
directory.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>