run-kvm: parse configs here too
It would be quite convenient if the run-kvm specific environment
variables could be included in the MMTests config file as well. Of
course, when the config file itself is then used by run-mmtests.sh
(either for traditional bare-metal run, or inside VMs), they will just
be ignored.

In order for this to be possible, we need to parse the config file(s) in
run-kvm.sh as well.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>