run-kvm and run-mmtests: synch tests execution in VMs
When running multiple tests in multiple VMs, one may want for each test
to start at the same time (actually, each iteration of each test, if the
test does more iterations), among the various VMs.

In order to be able to do so, the VMs must speak with a controller,
which will make sure the various phases of running the tests are
synchronized and proceed in lock-step.

In the simplest setup, the controller is the host. The VMs will send
tokens to the host as soon as they reach a particular 'state', and then
wait. The host will decide when it is the proper time to poke all the
VMs, signaling them that they can proceed in the following state.

This is, of course, all optional. In fact, this only happens if the
variable MMTESTS_HOST_IP is defined (in the environment or in a config
file). If it's not, run-mmtests.sh will be executed in all the VMs,
without any synchronization, as it happens curretnly, without this
patch.

For details about the state machine of the protocol used for exchanging
tokens between the VMs and the host, see the comment in run-kvm.sh.
For figuring out when VMs send the tokens for any give state,
check run-mmtests.sh and the shellpacks rewriting code.

Such oken exchanging happens currently over the network, via `nc`.
Most likely, this can be re-implemented using, for instance, something
like gRPC (either here, with https://github.com/fullstorydev/grpcurl)
or by putting together some service program.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>