run-kvm: fix quotes in the usage message
Due to using '`', virsh is invoked while displaying the help message,
which is...uhm, not ideal!

While there, also try to make it more clear that the parameters that
will be used for running `run-mmtests.sh` inside the VM(s) must be at
the end of the command line (of `run-kvm.sh`).

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>