fetch: Refetch files if the existing file size is 0
Currently mmtests use a fetched file if the file exists.
However such a existing file may be 0 bytes i.e last fetch failed to fetch
the file but left a 0 byte file.

Fix this by checking for the size of the existing file and not just for
the existence of the file.

Signed-off-by: Srikar Dronamraju <srikar@linux.vnet.ibm.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>