rewrite-shellpack: make argument parsing more robust
Throughout the codebase, there are many cases where we run thing like:
`my-shellpack.sh --foo $FOO --bar $BAR`

If the user accidentally forgot to set e.g. `$FOO` or `$BAR` in their
config file, then e.g. `my-shellpack.sh` is actually just passed
"--foo --bar".

The result is:

1. We treat "--bar" as the *value* of "--foo".
2. If there's an odd number, i.e., we have end up parsing a last flag
   with no value, then "shift 2" does nothing, and we infinite loop
   instead.

Instead of all this, try to detect missing values, and in that case
print out a helpful error message and exit with an error code.

Note that the new logic is still somewhat fragile (e.g. it assumes
real values never start with "-"). I think it would be better to
replace all of this with getopt(1) instead, but that's a bigger
rewrite and perhaps requires more discussion.

Signed-off-by: Axel Rasmussen <axelrasmussen@google.com>
Signed-off-by: mgorman <mgorman@suse.com>