sysbench: fix mariadb start argument plumbing
When parsing shellpack command line arguments, we want to be able to
detect when an argument was specified without an associated value. So,
we check if the next argument starts with a "-" (i.e., it's another
flag, not a value).

But, for mariadb we want to pass in *its flags as a value*, creating an
ambiguity. To resolve this situation, pass the arguments *for mariadb*
between *our scripts* with literal quotes. When we're ready to actually
pass them into mariadb *itself*, strip the quotes.

Signed-off-by: Axel Rasmussen <axelrasmussen@google.com>
Signed-off-by: mgorman <mgorman@suse.com>