pgbench: echo the logfile path to make debugging failures easier
If postgresql fails to start, pg_ctl unfortunately does not detect this.
It just forks postgresql, and exits immediately with a return code of 0.
So, if e.g. our configuration file is invalid, we just wait for a long
time for the pidfile to appear, and when this fails it's unclear how to
debug the issue.

It would be best if pg_ctl could just actually wait to see if postgres
immediately exits with an error, but I don't see a flag to make this
happen.

So, the least we can do is print out the logfile path, so the user needs
to do much less digging to figure out why things failed if the pidfile
never appears.

Signed-off-by: Axel Rasmussen <axelrasmussen@google.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>