fetch: show progress while downloading files
Say we have some install script which tries to fetch a thing, but for
whatever reason the download is very slow. Without this patch, we get
output like this:

whatever-install: Fetching from mirror http://foo.com/bar

And then we hang, perhaps for a very long time depending on the download
size / speed. To the user, it's unclear why we're hanging - is the
mirror just super slow? Or did we finish fetching the tarball, and then
hang at some other point?

Signed-off-by: Axel Rasmussen <axelrasmussen@google.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>