compare-kernels: make html header/footer work with --result-dir
--result-dir makes compare-kernels cd near the beginning of the script.

This can make the path in SCRIPTDIR invalid if it's relative, as it is
when compare-kernels is run from the root of mmtests.

The commands that print the html header and footer fail, and since the
one that prints the footer is the last in the script, compare-kernels
exits with 1.  Not immediately obvious why because the command's stderr
is redirected.

Just make SCRIPTDIR contain an absolute path.

Signed-off-by: Daniel Jordan <daniel.m.jordan@oracle.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>