run-kvm: collect output from run-mmtests.sh in VMs, via pssh
pssh can log the output of all the commands being executed
remotely, if an output directory where to put such log files is
provided. This may or may not be needed/useful in production, but
is definitely useful for development and debugging.

We also move the handling of the $_NCV variable, for providing
additional parameters to `nc`, in run-kvm.sh, as `nc` is invoked
from there too. This is really only intended as a debugging
facility.

Still related to `nc`, let's also make sure we kill it and clean
up after it only if we actually did start it. This should fix an
(harmless but annoying) error message being printed by `kill`
when we don't do VM synchronized runs.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>