run-kvm: allow the list of VMs to be specified in config file
It would be useful, especially useful now that we support host config
files, to be able to specify the list of VMs in other ways than just
via command line.

That may be sensibly more convenient than using a (very!) long list of
names on the command line, in the `--vm` parameter.

If both, CLI and config variable, are there the former has priority.

While there, make the CLI accepts `--vms` too.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>