Add support for tuned
tuned can be used to set a number of properties and apply a set of
tunings, according to some profiles. Add support for it, so that such
profile can be applied before starting a test.

As it is for `numad`, functions are in common files and both run-kvm.sh
and run-mmtests.sh use them. It is therefore possible to use a different
profile for the host and inside the guests.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>