run-kvm: support monitors at the host level
When running benchmarks inside VMs, it may be interesting to have the
monitors running not only inside the VMs themselves, but also on the
host. To catch, for example, anything related to virtualization itself,
capture total host load, etc.

In order to do so, move the handling of the monitors in functions in a
common file, that can be included by both run-kvm.sh and run-mmtests.sh.
There was a "monitoring related file", i.e., deferred-monitors.sh, so
put all the code for monitoring there (and rename it).

Note that, now that we have support for host specific config file, we
can have different monitors active on the host, with respects to the
ones that we have inside the guests.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>