install-depends: use command, as which itself may be missing
Seen this in a JeOS image. `which` is not there, so we need to install
it. But we use `which` to check what package manager to use, and the
whole thing fails quite spectacularly!

Use the `command` bash builtin for checking the package manager instead.
And add `which` as a package that needs to be there/needs to be
installed.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>