Visualize: scale load by num of CPUs for per-node heatmap
When building per-node heatmap, scale the per-node load measured
by mpstat by the number of CPUs we have on the nodes themselves.

If we don't do that, the graph often ends up empty, as the load
value used to index the color gradient array is too big (especially
when under high load).

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>