gitcheckout: add a last resort fallback to git.kernel.org
The config we have in configs/ uses $WEBROOT (and $SSHROOT)
which by default point to http://mcp/mmtests-mirror and that
is not available outside of SUSE infra. Of course one can
change that easily enough, but it's equally easy for us to
provide a fallback, in case the mirror is not there. This
way, we can tell users to just run:

  ./run-mmtests -c configs/config-workload-kerndevel TEST0

without having to also ask them to tweak the config file.

Note that I know we have the git_fetch shellpack command
and I tried using it. However, that fetch the repository
in work/sources rather than where we need then (in
work/testdisk/data) and moving the directory afterwords
looked clumsy.

Reported-by: Emilio Bruno <emilio.bruno@edu.unito.it>
Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>