bin: compare-mmtests.pl: add load of tests from JSON format
Enable the import of tests data contained in a JSON file
without the need for any log file.

This commit adds an option "--from-json" that takes a JSON
file with the following format:
[
    {
        "_ModuleName": "<test_type_name>",
        "_TestName": "<test_run_name>",
        "_ResultData": {
            "<operation>": [
                "Values": [ ],
                "SampleNrs": [ ]
            ]
        }
    }
]
Each element of the array is a test result and it will be loaded
as the test type indicated by the "_ModuleName" field.
Inside the "_ResultData" field are present the raw results of the test.
The "operation" field represents the operation type that the metrics
refer to, in sysbenchcpu's case is the number of threads used whereas
for iozone it can be the type of read or write done.
This JSON format is compatible with the output of the "--print-json"
option of the extract-mmtests.pl script.

With this workflow the work/log directory is not needed anymore to load
the tests data. This should work for, virtually, any test type.

Signed-off-by: Mirco Romagnoli <romagnoli.mirco@gmail.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>