bin: reports-from-json.pl: return benchmark type list from JSON
Read a JSON file and return a list of distinct test types.

With this commit the list of test types can be extracted from a JSON
file containing the modules extracted by the bin/extract-mmtests.pl
script with the "--print-json" option enabled.
The JSON file should have this structure (additional fields will be
ignored):
[
    {
        "_ModuleName": "<test_type_name>",
        "_TestName": "<test_run_name>",
        "_ResultData": {
            "<operation>": [
                "Values": [ ],
                "SampleNrs": [ ]
            ]
        }
    }
]

Signed-off-by: Mirco Romagnoli <romagnoli.mirco@gmail.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>