compare-kernels: add new flag "--from-json" to compare-kernels.sh
Enable the comparison of tests data from a JSON file.

The output will be the same as the previous version except that the
monitors won't be included in the output. This limitation will be
addressed in a later commit, if needed.

A JSON file extracted with the --print-json option of the
bin/extract-mmtests.pl script can be loaded by doing:
        ./compare-kernels.sh --from-json ./benchmarks.json

With this workflow the work/log directory is not needed anymore to load
the tests data. This should work for, virtually, any test type.

Signed-off-by: Mirco Romagnoli <romagnoli.mirco@gmail.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>