run-kvm: read the IPs of (already running) VMs from host config file
Make it possible to specify, in the host config file, a list of IPs.
They must correspond to already running VMs, and the list of their
names should be speficied as well.

If both are found, do not try to determine the IPs of the VMs from
the names and do not try to start (or stop, at the end of the benchmark)
them.

This makes it possible to run benchmarks inside VMs that we want or
need to manage (e.g., the power cycling) manually, instead than
automatically, for whatever reason.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>