mmtests: add test cmd line in json file
* Change extract script, extract-mmtests.pl
Add the check if command.log file exists in the result folder. When this
file is found, it is added to a json file for saving the commands that
are launched to perform the tests.

* Change extract class, Extract.pm
Add in Extract class the _Cmd attribute used to write in json file
the commands launched by the test.

* Add libfile-slurp-perl dependency
Request the dependency to be installed in run-mmtests.sh, and map the
correct package to install in install-depends

Signed-off-by: Luca Bartoli <228618@studenti.unimore.it>
Signed-off-by: Federico Gelmetti <federico.gelmo@gmail.com>