Add saladfork benchmark, local v. remote fork latencies
Forks on the remote NUMA node, i.e. when the parent and child are on different
nodes, are more expensive than local forks: they have higher latency. It could
be that a benchmark such as libmicro or lmbench reports a regression on the
fork(2) syscall latency, when the actual cause of the change is that the
scheduler has changed its preference on where to place new tasks.

This benchmark, called "saladfork", is specifically designed to detect when
that happens.

1. DESCRIPTION
2. INVOCATION AND REPORT
3. RAW DATA
4. CAVEATS
5. TODO

1. DESCRIPTION
--------------

The benchmark forks a child, the parent waits for the child to exit, the child
exits immediately. This is repeated SALADFORK_ITERATIONS times in a loop; the
latency of fork(2) is measured, and the CPUs of parent and child are
recorded.

Three "operations" (MMTests terminology) are reported:

* average latency of a local fork
* average latency of a remote fork
* share of local forks versus the total

2. INVOCATION AND REPORT
------------------------

A sample invocation follows.

  # ./run-mmtests.sh -c configs/config-scheduler-saladfork FOO
  # ./run-mmtests.sh -c configs/config-scheduler-saladfork BAR

The above execute two runs of saladfork, called FOO and BAR, so that we can
see what a comparison looks like.

    # ./bin/compare-mmtests.pl -d work/log -n FOO,BAR -b saladfork
                                                FO                     BA
                                               FOO                    BAR
    Min       local              106.00 (   0.00%)      122.00 ( -15.09%)
    Min       remote             126.00 (   0.00%)      148.00 ( -17.46%)
    Min       local_v_total        0.98 (   0.00%)        0.98 (  -0.01%)
    Amean     local              114.10 (   0.00%)      132.26 * -15.92%*
    Amean     remote             178.04 (   0.00%)      211.09 * -18.56%*
    Amean     local_v_total        0.98 (   0.00%)        0.98 *  -0.01%*
    Stddev    local               17.10 (   0.00%)       17.74 (  -3.72%)
    Stddev    remote              51.74 (   0.00%)       74.09 ( -43.20%)
    Stddev    local_v_total        0.00 (   0.00%)        0.00 (   0.00%)
    CoeffVar  local               14.99 (   0.00%)       13.41 (  10.52%)
    CoeffVar  remote              29.06 (   0.00%)       35.10 ( -20.78%)
    CoeffVar  local_v_total        0.00 (   0.00%)        0.00 (   0.00%)
    Max       local             1655.00 (   0.00%)     1782.00 (  -7.67%)
    Max       remote             504.00 (   0.00%)     1950.00 (-286.90%)
    Max       local_v_total        0.98 (   0.00%)        0.98 (  -0.01%)
    BAmean-50 local              110.21 (   0.00%)      127.04 ( -15.27%)
    BAmean-50 remote             141.01 (   0.00%)      169.17 ( -19.97%)
    BAmean-50 local_v_total        0.98 (   0.00%)        0.98 (  -0.01%)
    BAmean-95 local              112.02 (   0.00%)      129.92 ( -15.98%)
    BAmean-95 remote             170.75 (   0.00%)      201.92 ( -18.26%)
    BAmean-95 local_v_total        0.98 (   0.00%)        0.98 (  -0.01%)
    BAmean-99 local              113.19 (   0.00%)      131.33 ( -16.03%)
    BAmean-99 remote             176.07 (   0.00%)      207.42 ( -17.80%)
    BAmean-99 local_v_total        0.98 (   0.00%)        0.98 (  -0.01%)

Above we see that the arithmetic mean (Amean) of local fork latencies is
114.10 usec in the FOO run and 132.26 usec in the BAR run. Since we executed
FOO and BAR on the same system, the percentage difference between the two is
almost identical to the coefficient of variation (CoeffVar) of each run.
Similarly for remote forks: 178.04 usec +- 29.06% on FOO,
211.09 usec +- 35.10% on BAR.

What we really want to see here is the ratio of local forks versus the total,
which is 0.98: 98% of forks were local on both runs.

3. RAW DATA
-----------

The raw data reported by the benchmark looks like the following table:

    # zcat work/log/FOO/iter-0/saladfork/logs/saladfork-0.log.gz
    PARENT  CHILD   LATENCY
    022     023     163
    022     023     124
    022     023     131
    022     023     118
    022     023     124
    022     023     124
    022     023     131
    022     023     120
    022     023     118
    ...

The first line means that in the first iteration the parent was on CPU 22, the
child ended up in CPU 23, and fork() took 163 usec. There is a line for each
iteration:

    # zcat work/log/FOO/iter-0/saladfork/logs/saladfork-0.log.gz | wc -l
    100001

since the config sets SALADFORK_ITERATIONS=100000, plus one line for the table
header. The table is then processed by ExtractSaladfork.pm using the output of
numactl.

4. CAVEATS
----------

This benchmark's operation are very heterogeneous in nature: "local" and
"remote" are latencies, "local_v_total" is a unitless ratio. When summarizing
the result of a comparison in a single numeric score, the result may or may
not make sense. Marvin will look at the GMEAN value for the score (number in
the dashboard cell), and at DMEAN for its statistical significance (color of
the cell). Those are calculated when giving --print-ratio to compare-mmtests.pl.
From the above example:

    # ./bin/compare-mmtests.pl -d work/log -n FOO,BAR -b saladfork --print-ratio
                                            FO                     BA
                                           FOO                    BAR
    Ratio local                1.00 (0.00%) (+0.00s)        1.16 (-15.92%) (+1.04s)
    Ratio remote               1.00 (0.00%) (+0.00s)        1.19 (-18.56%) (+0.52s)
    Ratio local_v_total        1.00 (0.00%) (+0.00s)        1.00 (-0.01%) (+0.00s)
    Dmean Lower                0.00                         0.52
    Dmin  Lower                0.00                         0.00
    Dmax  Lower                0.00                         1.04
    Gmean Lower                1.00                         1.11

Here we have a GMEAN of 1.11, which for a lower-is-better benchmark like this
one, looks like a 11% regression. We've established already these results are
neutral, since FOO and BAR come from the same system. DMEAN is 0.52, which
says the FOO and BAR runs differ by about half a standard deviation, so the
dashboard wouldn't flag this as a regression -- in this specific example, all
is good.

5. TODO
-------

1. It would probably make more sense to count remote_v_total instead of
   local_v_total, so that all three operations are of the lower-is-better
   type. Local forks aren't inherently better of worse than remote ones, but
   if I had to pick, I'd favor local forks.

2. There is no data plotting support for this benchmark yet.

Signed-off-by: Giovanni Gherdovich <ggherdovich@suse.cz>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>