futexbench: Add wake-parallel workload
Unlike its 'wake' counterpart, the purpose of this is to
measure avg per-thread wakeup latencies when multiple
threads are doing 1/n wakeups. Units of measurement are
usec and can be influenced by hw artifacts and/or scheduling
decisions, so numbers should be taken with a grain of salt
along with standard deviation for actually determining
an issue. It is useful for detecting large/consistent spikes
in latency and determining consistently unusual numbers.

Signed-off-by: Davidlohr Bueso <dbueso@suse.de>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>