list-cpu-siblings.pl: correct cross-socket option on multi-node CPUs
On machines where the number of NUMA nodes equals the number of CPU
sockets, the cross-socket option selects CPUs on the same NUMA node and
the selected CPUs also usually share last-level cache. On machines where
the number of NUMA nodes is greater than the number of CPU sockets, the
cross-socket option may cause mmtests to select CPUs that are not located
on the same NUMA node. This causes dyscrepancies in performance results.

This patch seeks to unify the effects of the cross-socket option on
different machines by introducing a new set of CPUs - CPUs that share an
L3 cache. There is a fallback to node_cores CPUs in cases where the kernel
does not provide information about L3 cache (e.g. older kernels running on
aarch64 CPUs) or all the other CPUs sharing an L3 cache are siblings of
the specified CPU (e.g. a ppc64le machine with 8 SMT threads per core and
each core has its own L3 cache).

This patch also fixes a bug that caused SMT siblings to be included in the
sibling and node_core lists when the cpulist and core_siblings_list files
do not contain ranges.

Signed-off-by: Jiri Wiesner <jwiesner@suse.de>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>