Limit fio randread tests to 5 minutes
Currently config-io-fio-randread-doublemem-4k-4t and
config-io-fio-randread-doublemem-32k-4t tests run file size to 2 times
the memory. For rotational storage doing 128GB worth of random reads
takes ~40 hours and it is mostly a waste of time. Limit the runtime of
the tests to 5 minutes. That should be more than enough to get idea how
random reads perform.

Signed-off-by: Jan Kara <jack@suse.cz>