run-kvm: don't explode if '-k' is not used, and fix config handling
The default behavior used to be to install the current host kernel
inside of the VM, and then boot it. This is fine, but is currently not
implemented properly, if we're running outside of Marvin and/or if we
are using multiple VMs.

This means that the '-k' parameter is pretty much mandatory, in such
cases (or run-kvm.sh will just crash). Improve this, by checking if the
kvm-boot script, which is required for using the host kernel, is there
and, if not, just behave like if '-k' were passed, even if that is not
the case.

While there, also fix the issue that, even if specifying (one or more)
guest config file(s), only the default one (i.e., 'config' inside
$SCRIPTDIR) was being used.

Signed-of-by: Dario Faggioli <dfaggioli@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>