run-kvm: take care of the firewall, for VMs-host communications
When running benchmarks in multiple VMs, they need to communicate with
the host for coordination. If this happens over the network (which is
the only supported way, right now) we must be sure that the firewall is
not blocking  it, at either side.

There could be many firewalling technology and setups in place, both on
the host and in the guest, and we probably will never be able to
automatically take care of all of them. Let's, however, at least try to
deal with some simple situations, mainly if firewalld is being used and
firewall-cmd is available.

If that does not work, the recommendation remains to disable the
firewall before starting the benchmarks.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>