kvm-ip-address: deal with miltiple entries
It's possible that a VM has multiple NICs, so deal with that. It's also
possible that `ip n show` report multiple reachable IP addresses (if we
check it before it manages to mark them as STALE).

The policy, for now, is to return the first IP that answers to a quick
ping.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>