run-kvm.sh: don't depend (too much) on GNU parallel
Like pscp, GNU parallel may not be part of the distribution. If it is,
we want to use it, for trying to send the tokens that triggers the
various phases of the tests to the VMs all at the same time. However,
if it's not there, we can hopefullly achieve a similar enough behavior
via background execution.

So, if the tool is not there, go with that, instead of exiting.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>