run-kvm.sh: fix handling -m and -n and don't copy archives
The -m and -n options are accepted by both run-kvm.sh and
run-mmtests.sh. This means that, when run-kvm.sh is used, the can appear
in the command line twice. In which case, it should "consume" just the
first occurrence of them, and leave the other one in there for when
run-mmtests.sh will be executed in the guest(s). This is basically the
same that we're doing already for -p.

While there, keep any file with named work* out of the archive that we
copy into the guest(s). In fact, after a couple of execution, we may have
a few archives whose names start with "work" in the main MMTests
directory, and there's no point copying them into the guest(s).

Reported-by: Lorenzo Dentis <lorenzo.dentis@edu.unito.it>
Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>