config-io-fio-sync-maxrandwrite: Make workload avoid repetitions
Avoid time_based directive to fio since if we manage to write the whole
file (for fast storage and / or small filesystems) we will just loop
over the same file again which is already in cache thus skewing results.

Also drop pointless iodepth directive. It does nothing for buffered IO.

Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>