config-io-fio-randread-sync-heavywrite: Avoid rereading files
The results of the workload (especially for reads) are frequently skewed
because the jobfile is set to run for 5 minutes and if during that time
we manage to read all files, the readers loop, reading from page cache
now. Modify the job file so that writers write in a loop for 5 minutes
and the job terminates whenever either reader reads all files or writers
terminate.

Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>