io_uring: Introduce io_uring nops test
A common way to validate io_uring performance, used frequently on the
mailing list, is to count the number of NOP commands completed in a
period of time.  A tool for this is available in the fio repo, commonly
referred as  t/io_uring.

This patch introduces a shellpack to execute this test, by leveraging
the existing fio repository installation infrastructure.  It introduces
a new test type called io_uring for this tool.  A following commit will
introduce the infrastructure to parse the results.

Signed-off-by: Gabriel Krisman Bertazi <krisman@suse.de>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>