perfnuma-bench: Don't use '-0' option to avoid OOM killer
With PERFNUMA_PROCESS_WSS=$((${MMTESTS_NODE_SIZE[0]}*4/5)) as set in
configs/config-numa-perfnuma, OOM killer might be invoked on
multi-node systems because memory allocations are restricted to node
0.
For more than one numa-mem process this amount of memory is mmaped and
accessed on node 0. And eventually OOM killer kills one of the
processes. This in turn causes endless stall in perf-bench-numa. It
waits for creation of tasks that will never happen because a process
that would do this was killed.

Remove '-0' from 'perf bench numa' invocation and allow allocation from
different numa nodes.

Signed-off-by: Andreas Herrmann <aherrmann@suse.de>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>