gitsource-bench: Skip tests with prereq to AUTOIDENT or !AUTOIDENT
It depends on the environment/config which tests are executed. It's
unreliable for automated performance test execution (e.g. when
comparing native with container environment).
To reliably execute same set of tests just skip the small number of
tests with prereq of [!]AUTOIDENT.

Affected tests are #3 in t0007-git-var.sh, #34 and #35 in
t7502-commit.sh, #26 and #27 in t9001-send-email.sh.

Signed-off-by: Andreas Herrmann <aherrmann@suse.de>