run-container: Ensure large enough pids-limit for container
Option 'pids-limit=-1' might not set pids.max to 'max'.
It depends on default configuration for DefaultTasksMax.
If not set to 'infinity', pids.max is limited by the specified value.
For some tests this limit might be too low and might prevent creation
of required number of threads.
Set large pid limit (15000) in case of DefaultTasksMax!=infinity.

Signed-off-by: Andreas Herrmann <aherrmann@suse.de>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>