gnupgsource: Add new shellpack
This is running 'make check' of gnupg build as a benchmark.
(This is kind of similar to gitsource benchmark -- using a unit test
of a software package as workload.)

I've picked version 'gnupg-2.2.27' -- this matched gnupg version
installed on a current openSUSE Leap 15.4 system where I've tested
this.

I didn't chose a more recent version of gnupg because this requires
newer versions of some build tools and/or libraries.

Signed-off-by: Andreas Herrmann <aherrmann@suse.de>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>