profile-hooks-perf: generate many reports by default
The perf post hook has been reworked to detect the common types of perf
data files - a stat file, a profile and a profile with call graphs.
Detailed reports are generated for each of the common types of perf data
files. Stat data file reports contain total counts and per CPU counts.
Profile reports are generated for DSOs, DSOs-symbols and
DSOs-symbols-CPUs. Call graph reports are generated in different variants,
all in extra detail, and also without call graphs to get an approximate
profile. The rationale for generating multiple reports is that they cover
the basic questions one needs to answer and the reports tend to be smaller
than the original data files (hence better for archiving). Optionally, the
perf report command may be specified to disable the creation of default
reports.

Signed-off-by: Jiri Wiesner <jwiesner@suse.de>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>