netperf: add ipv6 support
The version of the IP protocol is defined in driver scripts, which is
consistent with the transport layer protocol being defined in driver
scripts. This lead to the need to have a separate driver script for every
combination of a network layer protocol and a transport layer protocol.
The configurations and perl extraction scripts mirror this arrangement.
Configurations without an explicit network layer protocol version were
declared obsolete but kept for the time being to retain compatibility with
existing results.

Signed-off-by: Jiri Wiesner <jwiesner@suse.de>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>