fio_bench: merge block and jobfile handling
There is nothing special about block devices, other than the fact that
we commonly use a fixed file target for them instead of letting fio
decide on a file name.  Provide a script option to specify the target
instead and rely on that, such that filesystem tests sharing a file can
also use the filename argument.

Signed-off-by: Gabriel Krisman Bertazi <krisman@suse.de>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>