fio-bench: flexibilize non-jobs testing
Add a new flavor of fio tests that generically scales any, single or
multiple variables on the same test.  The reasoning is that we have a
few flavors that scale njobs, rw, and size already, and I wanted to add
tests that vary queue depth for io_uring.  So, instead of a new flavor,
introduce this more generic infrastructure.  Further patches will
reimplement the existing tests on top of it.

Fortunately, we can leverage the existing parsing infrastructure without
modifications.  We create unique names for each subtest, which will be
accumulated separately in compare-kernels.

I'm using this on the following patches that add more io_uring testing.

Signed-off-by: Gabriel Krisman Bertazi <krisman@suse.de>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>