configs: Add configuration for io depth scaling test
This test submits batched IO with fio, scaling up the io submission
depth on a single cpu.  On a fast storage device, with nr_requests>>fio
depth, I'm observing cpu saturation around qd128, so go beyond that.  it
is also interesting to observe a comparison of libaio and io_uring on
the same device.

Signed-off-by: Gabriel Krisman Bertazi <krisman@suse.de>
Signed-off-by: Mel Gorman <mgorman@techsingularity.net>