MMTests::Stat: replace List::BinarySearch
List::BinarySearch is ancient and has a few issues:

  a) The project hasnt't received any updates since 2018
  b) It's licensed under GPLv1 or Artistica v1 license which are
     not compatible with GPLv2
  c) Some distributions don't carry a package for it

Debian in particular doesn't have it anymore. I looked at the code
and was bewildered with the amount of insane hacks to support ancient
versions of perl and to also support running our custom comparison
routine.

All the above challenges seem silly to deal with on users if we can
just implement what we need and call it a day. So let's do that.
Instead of piling a set of unit tests on mmtests the approach taken
to ensure a proper replacement for the library is to use a separate
repository that proves a replacement for the library, which has its
own unit tests [0]. The work we do here is then to not only show how
we have simplified the old code, but also can use the same tests in
the kill-List-BinarySearch against the exact old library without
breaking things. That gives us two ways to test for correctness.

While at it use SPDX license identifiers to simplify license clarity
just as we have in the kernel. The same SPDX tag is already upstream
on Linux. We can expand on this to keep things simpler later.

[0] https://github.com/mcgrof/kill-List-BinarySearch

Reviewed-by: Jan Kara <jack@suse.cz>
Tested-by: Davidlohr Bueso <dave@stgolabs.net> [futex, hackbench, sysvs ipc]
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>